ref(pool): apply changes required from code review
