+++ api_pool_test.go
@@ -2,8 +2,10 @@ package blockfrost_test
 
 import (
 	"context"
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"path/filepath"
 	"testing"
 	"time"
 
@@ -17,13 +19,13 @@ func TestResourcePools(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					"pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy",
-					"pool1hn7hlwrschqykupwwrtdfkvt2u4uaxvsgxyh6z63703p2knj288",
-					"pool1ztjyjfsh432eqetadf82uwuxklh28xc85zcphpwq6mmezavzad2"
-				]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
+
 		}),
 	)
 	defer s.Close()
@@ -50,18 +52,12 @@ func TestResourcePoolsRetired(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-						"pool_id": "pool19u64770wqp6s95gkajc8udheske5e6ljmpq33awxk326zjaza0q",
-						"epoch": 225
-					},
-					{
-						"pool_id": "pool1dvla4zq98hpvacv20snndupjrqhuc79zl6gjap565nku6et5zdx",
-						"epoch": 215
-					}
-				]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_retired.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -88,18 +84,12 @@ func TestResourcePoolsRetiring(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-						"pool_id": "pool19u64770wqp6s95gkajc8udheske5e6ljmpq33awxk326zjaza0q",
-						"epoch": 225
-					},
-					{
-						"pool_id": "pool1dvla4zq98hpvacv20snndupjrqhuc79zl6gjap565nku6et5zdx",
-						"epoch": 215
-					}
-				]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_retiring.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -120,31 +110,39 @@ func TestResourcePoolsRetiring(t *testing.T) {
 func TestResourcePoolSpecific(t *testing.T) {
 	t.Parallel()
 	inputPoolID := "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy"
-	expectedElement := blockfrost.PoolHistory{
-		Epoch:           233,
-		Blocks:          22,
-		ActiveStake:     "20485965693569",
-		ActiveSize:      1.2345,
-		DelegatorsCount: 115,
-		Rewards:         "206936253674159",
-		Fees:            "1290968354",
+	expectedElement := blockfrost.PoolSpecific{
+		PoolID:         inputPoolID,
+		Hex:            "0f292fcaa02b8b2f9b3c8f9fd8e0bb21abedb692a6d5058df3ef2735",
+		VrfKey:         "0b5245f9934ec2151116fb8ec00f35fd00e0aa3b075c4ed12cce440f999d8233",
+		BlocksMinted:   69,
+		LiveStake:      "6900000000",
+		LiveSize:       0.42,
+		LiveSaturation: 0.93,
+		LiveDelegators: 127,
+		ActiveStake:    "4200000000",
+		ActiveSize:     0.43,
+		DeclaredPledge: "5000000000",
+		LivePledge:     "5000000001",
+		MarginCost:     0.05,
+		FixedCost:      "340000000",
+		RewardAccount:  "stake1uxkptsa4lkr55jleztw43t37vgdn88l6ghclfwuxld2eykgpgvg3f",
+		Owners:         []string{"stake1u98nnlkvkk23vtvf9273uq7cph5ww6u2yq2389psuqet90sv4xv9v"},
+		Registration: []string{
+			"9f83e5484f543e05b52e99988272a31da373f3aab4c064c76db96643a355d9dc",
+			"7ce3b8c433bf401a190d58c8c483d8e3564dfd29ae8633c8b1b3e6c814403e95",
+			"3e6e1200ce92977c3fe5996bd4d7d7e192bcb7e231bc762f9f240c76766535b9",
+		},
+		Retirement: []string{"252f622976d39e646815db75a77289cf16df4ad2b287dd8e3a889ce14c13d1a8"},
 	}
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					  "epoch": 233,
-					  "blocks": 22,
-					  "active_stake": "20485965693569",
-					  "active_size": 1.2345,
-					  "delegators_count": 115,
-					  "rewards": "206936253674159",
-					  "fees": "1290968354"
-					}
-				  ]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_specific.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -155,11 +153,11 @@ func TestResourcePoolSpecific(t *testing.T) {
 		t.Fatalf(err.Error())
 	}
 	q := blockfrost.APIPagingParams{}
-	pool, err := api.PoolHistory(context.TODO(), inputPoolID, q)
+	pool, err := api.PoolSpecific(context.TODO(), inputPoolID, q)
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
-	assert.Equal(t, expectedElement, pool[0])
+	assert.Equal(t, expectedElement, pool)
 }
 
 func TestResourcePoolHistory(t *testing.T) {
@@ -177,19 +175,12 @@ func TestResourcePoolHistory(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					  "epoch": 233,
-					  "blocks": 22,
-					  "active_stake": "20485965693569",
-					  "active_size": 1.2345,
-					  "delegators_count": 115,
-					  "rewards": "206936253674159",
-					  "fees": "1290968354"
-					}
-				  ]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_history.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -223,18 +214,12 @@ func TestResourcePoolMetadata(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`{
-					"pool_id": "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy",
-					"hex": "0f292fcaa02b8b2f9b3c8f9fd8e0bb21abedb692a6d5058df3ef2735",
-					"url": "https://stakenuts.com/mainnet.json",
-					"hash": "47c0c68cb57f4a5b4a87bad896fc274678e7aea98e200fa14a1cb40c0cab1d8c",
-					"ticker": "NUTS",
-					"name": "Stake Nuts",
-					"description": "The best pool ever",
-					"homepage": "https://stakentus.com/"
-				  }`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_metadata.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -264,17 +249,12 @@ func TestResourcePoolRelay(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					  "ipv4": "4.4.4.4",
-					  "ipv6": "https://stakenuts.com/mainnet.json",
-					  "dns": "relay1.stakenuts.com",
-					  "dns_srv": "_relays._tcp.relays.stakenuts.com",
-					  "port": 3001
-					}
-				  ]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_relays.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -301,22 +281,12 @@ func TestResourcePoolDelegator(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					  "address": "stake1ux4vspfvwuus9uwyp5p3f0ky7a30jq5j80jxse0fr7pa56sgn8kha",
-					  "live_stake": "1137959159981411"
-					},
-					{
-					  "address": "stake1uylayej7esmarzd4mk4aru37zh9yz0luj3g9fsvgpfaxulq564r5u",
-					  "live_stake": "16958865648"
-					},
-					{
-					  "address": "stake1u8lr2pnrgf8f7vrs9lt79hc3sxm8s2w4rwvgpncks3axx6q93d4ck",
-					  "live_stake": "18605647"
-					}
-				  ]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_delegator.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -346,14 +316,12 @@ func TestResourcePoolBlocks(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					"d8982ca42cfe76b747cc681d35d671050a9e41e9cfe26573eb214e94fe6ff21d",
-					"026436c539e2ce84c7f77ffe669f4e4bbbb3b9c53512e5857dcba8bb0b4e9a8c",
-					"bcc8487f419b8c668a18ea2120822a05df6dfe1de1f0fac3feba88cf760f303c",
-					"86bf7b4a274e0f8ec9816171667c1b4a0cfc661dc21563f271acea9482b62df7"
-				  ]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_block.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -382,25 +350,12 @@ func TestResourcePoolUpdate(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					  "tx_hash": "6804edf9712d2b619edb6ac86861fe93a730693183a262b165fcc1ba1bc99cad",
-					  "cert_index": 0,
-					  "action": "registered"
-					},
-					{
-					  "tx_hash": "9c190bc1ac88b2ab0c05a82d7de8b71b67a9316377e865748a89d4426c0d3005",
-					  "cert_index": 0,
-					  "action": "deregistered"
-					},
-					{
-					  "tx_hash": "e14a75b0eb2625de7055f1f580d70426311b78e0d36dd695a6bdc96c7b3d80e0",
-					  "cert_index": 1,
-					  "action": "registered"
-					}
-				  ]`,
-			))
+			fp := filepath.Join(testdata, "json", "pool", "pools_update.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
